Working on the HTML/CSS for Main Page
I went to Layoutit!'s Website and downloaded a bootstrap code for the Navbar and the Carousel.  I customized the navbar some more by deleting fields we didn't need (like more buttons) and took out the search from the top.  On the Carousel,  I added some text that made sense.  My next step for that is to find better photos to display.  Third,  I found some code for a search bar.  I put it below the Carousel so people can scroll down and type in what they are looking for.  My next step is to customize it some more by changing the color of the nav bar, adding better photos, and perhaps making the Search Bar bigger.