Added more html pages
I set the nav-bar up so that all of the tabs will lead you to the correct page.
IE: About Us, Browse Directories, and Campuses (all four campuses).

I set up the URLS, the views, and the templates.  The templates are very plain-- it extends the base that has the nav bar, and each page has what will be going on that page.