Added some html and css.
About Us- I added a short description of the Ungmap page.  We can always change this.  I also added a section "About the Programmers" where I divided it up into three columns-- one for each of us.  I added a small 80x80 avatar for each programmer that takes the user to their github account on click.

Base.html-- Not a whole lot changed here.  I added a small footer.  I can change it at anytime...just thought I'd put one in so it didn't look so funny.

Main_page:  I started messing around with Haystack on the front page.  I took out the radio buttons that I had thinking that Crystal's haystack code will automatically populate them like on the search.html.  I changed the search button to a submit, so the form is looking for an input from the user.  Right now, if a user types something on the front page search, it'll just add a "/?" on the end of the html.  ...Still a work in progress
         In addition, I deleted two of the holder photos in the carousel to panoramic pictures of UNG I found online.  I wanted to put these in for our "show and tell."  We will need to take our own pictures soon.

 Navbar-- fixed the weird color that was going on when you clicked on the campus dropdown....it is now a white background with dark colored font.

 Models.py-- justed added the plural verbose that we all agreed on.