Added GeoLocation to Campus Maps
I went tot he campus_detail.html page and added in the code to work with the Geolocation function.  I used https://www.mapbox.com/mapbox.js/example/v1.0.0/geolocation/ as a guide with working with the GeoLocation.  When a user clicks the "Find Me" button on the top of the campus detail page, it will pop up a point of where they are at (this works really well with Mobile Devices with the GPS turned on).

I am also pushing up the loads with my file paths instead of Will's.  I also just added some Lorem Ipsom text in the About Us page.